Move constants into their own header
